requirements.txt
rolling-dice
aiogram
PyYAML